DR3_AT/requirements.txt
statsbombpy
fastapi
uvicorn
streamlit
langchain
openai
pydantic
requests
matplotlib
python-dotenv
